refactor: rename options to overrides
